Dodanie biblioteki pdf do studenta i korepetytora, dodanie customizac…

+++ backend/src/main/java/com/vectorpeaks/backend/repository/ReviewRepository.java
@@ -1,8 +1,8 @@
 /*
  * ReviewRepository.java
  *
- * Version: 1.1
- * Date: 2026-05-15
+ * Version: 1.2
+ * Date: 2026-05-24
  *
  * Copyright (c) 2026 EduLink Team. All rights reserved.
  *
@@ -21,19 +21,44 @@ import java.util.Optional;
 
 /**
  * Repository interface for performing CRUD operations on {@link Review} entities.
+ * Provides custom queries for average rating, review count, and lookup by booking/tutor.
  *
  * @version 1.1
  * @author EduLink Team
  */
 public interface ReviewRepository extends JpaRepository<Review, Integer> {
 
+    /**
+     * Retrieves the average rating for a given tutor.
+     *
+     * @param tutorId the ID of the tutor
+     * @return the average rating as a Double, or null if no reviews exist
+     */
     @Query("SELECT AVG(r.rating) FROM Review r WHERE r.tutorId = :tutorId")
     Double getAverageRatingByTutorId(@Param("tutorId") Integer tutorId);
 
+    /**
+     * Counts the number of reviews for a given tutor.
+     *
+     * @param tutorId the ID of the tutor
+     * @return the total number of reviews
+     */
     @Query("SELECT COUNT(r) FROM Review r WHERE r.tutorId = :tutorId")
     Integer countReviewsByTutorId(@Param("tutorId") Integer tutorId);
 
-    Optional<Review> findByBookingId(Long bookingId);
-    
+    /**
+     * Finds a review by the associated booking ID.
+     *
+     * @param bookingId the ID of the booking
+     * @return an Optional containing the review if found, or empty otherwise
+     */
+    Optional<Review> findByBookingId(Integer bookingId);
+
+    /**
+     * Retrieves all reviews for a given tutor.
+     *
+     * @param tutorId the ID of the tutor
+     * @return list of reviews belonging to the tutor
+     */
     List<Review> findByTutorId(Integer tutorId);
 }
